commet function

+++ app/Http/Controllers/Admin/UserController.php
@@ -1,13 +1,10 @@
 <?php
 
-
 namespace App\Http\Controllers\Admin;
 
 use App\Services\UserService;
 use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
 use App\Http\Requests\User\UserCreateFormRequest;
-use App\Models\User;
 use App\Http\Requests\User\UserUpdateFormRequest;
 
 class UserController extends Controller
@@ -18,32 +15,71 @@ class UserController extends Controller
     {
         $this->userService = $userService;
     }
+    /**
+     * Desc: Danh sách user
+     *
+     */
     public function index()
     {
         $lists = $this->userService->getAll();
         return view('admin.users.list', compact('lists'));
     }
+
+    /**
+     * Desc: Giao diện thêm user
+     *
+     */
     public function add()
     {
         return view('admin.users.add');
     }
+
+    /**
+     * Desc: Phương thức thêm user, thêm người dùng vào base
+     *
+     */
     public function postAdd(UserCreateFormRequest $request)
     {
         $result = $this->userService->createUser($request);
         return redirect()->route('admin.users.index')->with('msg', 'Thêm người dùng thành công');
     }
-    public function edit(User $user)
+
+    /**
+     * Desc: Giao diện sửa user
+     *
+     */
+    public function edit(string $user)
     {
-        return view('admin.users.edit', compact('user'));
+        $user = $this->userService->getOne($user);
+        if ($user) {
+            return view('admin.users.edit', compact('user'));
+        }
+        return redirect()->route('admin.users.index')->with('err', 'Có lỗi xảy ra, không thể sửa thông tin người dùng này');
     }
-    public function postEdit(User $user, UserUpdateFormRequest $request)
+    /**
+     * Desc: Phương thức cập nhật thông tin user trong database
+     *
+     */
+    public function postEdit(string $user, UserUpdateFormRequest $request)
     {
-        $result = $this->userService->updateUser($user, $request);
-        return back()->with('msg', 'Cập nhật người dùng thành công');
+        $user = $this->userService->getOne($user);
+        if ($user) {
+            $result = $this->userService->updateUser($user, $request);
+            return back()->with('msg', 'Cập nhật người dùng thành công');
+        }
+        return back()->with('err', 'Có lỗi xảy ra, không thể cập nhật thông tin người dùng này');
     }
-    public function delete(User $user)
+    /**
+     * Desc: Phương thức xóa user khỏi database
+     *
+     */
+    public function delete(string $user)
     {
-        $result = $this->userService->deleteUser($user);
-        return redirect()->route('admin.users.index')->with('msg', 'Xóa người dùng thành công');
+        $user = $this->userService->getOne($user);
+        if ($user) {
+            $result = $this->userService->deleteUser($user);
+            return redirect()->route('admin.users.index')->with('msg', 'Xóa người dùng thành công');
+        }
+        return redirect()->route('admin.users.index')->with('err', 'Có lỗi xảy ra, không thể xóa thông tin người dùng này');
     }
 }
